fix(filer): propagate isIgnored to nested entries
intent(filer): nested node_modules and other ignored entries should grey out like top-level
decision(filer): extract proto→FileEntry conversion as toFileEntries in filerUtils.ts as SSOT, route both FilerPane and FileTreeItem.loadChildren through it
rejected(filer): only swap the hardcoded isIgnored: false in FileTreeItem.loadChildren — leaves duplicated conversion in two files, same drop-out reappears next time a field is added
learned(filer): native FSOps.readDir already feeds nested relative paths to git check-ignore --stdin -z and returns correct isIgnored per entry; the bug lived entirely in the renderer-side child loader discarding the flag
