+++ apps/renderer/src/features/filer/filerUtils.ts
@@ -45,6 +45,15 @@ function getDeletedEntries(dirPath: string, gitStatuses: Record<string, string>)
   }));
 }
 
+/** proto の FsReadDirEntry を FileEntry に変換する */
+function toFileEntries(entries: { name: string; type: string; isIgnored: boolean }[]): FileEntry[] {
+  return entries.map((e) => ({
+    name: e.name,
+    isDirectory: e.type === "directory",
+    isIgnored: e.isIgnored,
+  }));
+}
+
 /** ディレクトリパスの末尾から表示名を抽出 */
 function dirName(dirPath: string): string {
   const parts = dirPath.split("/");
@@ -61,5 +70,5 @@ function sortEntries(entries: FileEntry[]): FileEntry[] {
   });
 }
 
-export { dirName, getDeletedEntries, sortEntries };
+export { dirName, getDeletedEntries, sortEntries, toFileEntries };
 export type { FileEntry };
